Python Exercise
This is advertising campaign engagement data from Facebook. Each row is an ad that ran on the platform. It includes the amount spent on the campaign, and other outcome/engagement data. For your first homework, let's do some analyses that tease out some insights on how to advertise more effectively.